requirements.txt
streamlit
nltk
spacy
textblob
scikit-learn
pandas
numpy
matplotlib
fpdf
supabase
python-dotenv